Use `rate.Limiter` with go routines
Serial wasn't cutting it, we need concurrency to reach the connection rates
I require. I initially tried it with a `Sometimes` but that seemed to have too
much overhead. They do state in the docs that calling `Do` at the same time
will cause the functions to be run in serial. Using `rate.Limiter` allows me
to get to at least 100 connections per second on a single box.